requirements.txt
pydobotplus
pydobot
pyserial
PyQt5
pyqtgraph>=0.13
numpy
